requirements.txt
# Core Framework
streamlit>=1.28.0
gradio>=4.0.0
# Pin compatible versions to avoid Gradio introspection errors
pydantic==2.10.6
fastapi<0.116
starlette<0.47  # Parallel Gradio implementation

# LLM Provider
openai>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Note: No FHIR parsing library needed - raw FHIR JSON passed directly to LLM

# Additional utilities (if needed)
# pydantic>=2.0.0  # For data validation
# typing-extensions>=4.8.0  # For type hints

